% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pubtator_sqlite_path.R
\name{make_pubtator_sqlite_path}
\alias{make_pubtator_sqlite_path}
\title{Make a path to the PubTator sqlite file.}
\usage{
make_pubtator_sqlite_path(pt_path)
}
\arguments{
\item{pt_path}{A character string indicating the full path of the
directory containing the pubtator gz files to be extracted.}
}
\value{
A character string indicating the full path to the sqlite file.
}
\description{
Make a path to the PubTator sqlite file.
}
