% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_columns.R
\name{pt_columns}
\alias{pt_columns}
\title{List the column names for a table in the PubTator sqlite database}
\usage{
pt_columns(db_con, table_name)
}
\arguments{
\item{db_con}{A connection to the PubTator sqlite database, as created via
pubator_connector.}

\item{table_name}{The name of the table of interest. Valid tables can be
found using pt_tables. Capitalization does not matter.}
}
\value{
A character vector of the column names for a given table.
}
\description{
List the column names for a table in the PubTator sqlite database
}
\examples{
\donttest{
db_con <- pt_connector(pt_path)
pubtator_columns(db_con, "gene")
}
}
