% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_connector.R
\name{pt_connector}
\alias{pt_connector}
\title{Connect to pubtator.sqlite}
\usage{
pt_connector(pt_path)
}
\arguments{
\item{pt_path}{A character string indicating the full path of the
directory containing the pubtator gz files to be extracted.}
}
\value{
A SQLiteConnection
}
\description{
Connect to pubtator.sqlite
}
\examples{
\donttest{
pt_connector("D:/Reference_data/PubTator")
}
}
