% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pt_to_sql.R
\name{pt_to_sql}
\alias{pt_to_sql}
\title{Create sqlite database from the pubtator data.}
\usage{
pt_to_sql(pt_path, skip_behavior = TRUE, remove_behavior = FALSE)
}
\arguments{
\item{pt_path}{A character string indicating the full path of the
directory containing the pubtator gz files to be extracted.}

\item{skip_behavior}{TRUE/FALSE indicating whether the file should be
re-extracted if it has already been extracted.}

\item{remove_behavior}{TRUE/FALSE indicating whether the gz files should be
removed following successful extraction.}
}
\description{
Create sqlite database from the pubtator data.
}
\examples{
\donttest{
download_path <- tempdir()
current_dir <- getwd()
setwd(download_path)
pt_to_sql("PubTator")
setwd(current_dir)
}
}
