% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubtator_citations.R
\name{pubtator_citations}
\alias{pubtator_citations}
\alias{pubtator_citation}
\title{See the citations for PubTator}
\usage{
pubtator_citations()
}
\description{
See the citations for PubTator
}
\examples{
pubtator_citations()
}
