% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubtator_ftp_url.R
\name{pubtator_ftp_url}
\alias{pubtator_ftp_url}
\title{NCBI's ftp url definition for PubTator.}
\usage{
pubtator_ftp_url()
}
\value{
A character string giving the ftp url for PubTator.
}
\description{
NCBI's ftp url definition for PubTator.
}
