% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubtator_tables.R
\name{pubtator_tables}
\alias{pubtator_tables}
\title{Table and dataset definitions}
\usage{
pubtator_tables()
}
\value{
A character vector where names are table names and values are dataset
  names.
}
\description{
Table and dataset definitions
}
