% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_pt.R
\name{download_pt}
\alias{download_pt}
\title{Download PubTator data via ftp.}
\usage{
download_pt(pubtator_parent_path, ...)
}
\arguments{
\item{pubtator_parent_path}{The path to the directory where the PubTator data
folder will be created.}

\item{...}{Additional arguments to dir.create and download.file.}
}
\value{
The path to the newly created directory. This can be passed to other
  functions as the pt_path argument.
}
\description{
Download PubTator data via ftp.
}
\examples{
\donttest{
# Use the full path. The files are large. Writing somewhere other than the
# temp directory is recommended.
download_path <- tempdir()
download_pt(dowload_path)
}
}
